+++ src/main/java/jardin/empresa/service/impl/ProyectoServiceImpl.java
@@ -1,6 +1,7 @@
 package jardin.empresa.service.impl;
 
 import jardin.empresa.DTO.ProyectoDTO;
+import jardin.empresa.exception.NotFoundException;
 import jardin.empresa.mapper.ProyectoMapper;
 import jardin.empresa.model.Proyecto;
 import jardin.empresa.repository.ProyectoRepository;
@@ -8,6 +9,7 @@ import jardin.empresa.service.ProyectoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -31,12 +33,19 @@ public class ProyectoServiceImpl implements ProyectoService {
     @Override
     public ProyectoDTO get(Long id) {
         Optional<Proyecto>proyecto = proyectoRepository.findById(id);
+        if(!proyecto.isPresent()){
+            throw new NotFoundException("No existe el proyecto: " + id);
+        }
         ProyectoDTO dto = proyectoMapper.entityToDto(proyecto.get());
         return dto;
     }
 
     @Override
     public void delete(Long id) {
+        Optional<Proyecto>proyecto = proyectoRepository.findById(id);
+        if(!proyecto.isPresent()){
+            throw new NotFoundException("No existe el proyecto: " + id);
+        }
         proyectoRepository.deleteById(id);
     }
 
@@ -48,5 +57,11 @@ public class ProyectoServiceImpl implements ProyectoService {
         return dto;
     }
 
+    @Override
+    public List<ProyectoDTO> getList() {
+        List<Proyecto> proyectos = proyectoRepository.findAll();
+        List<ProyectoDTO> proyectosDTO = proyectoMapper.listEntityDto(proyectos);
+        return proyectosDTO;
+    }
 
 }
